python_module_8/ex2/pyproject.toml
[project]
name = "ex2"
version = "0.1.0"
description = ""
authors = [
    {name = "thribeir",email = "thiagoribeirochico@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"